+++ src/eml2_3/TimeSeries.cpp
@@ -51,7 +51,7 @@ void TimeSeries::pushBackTimestamp(const tm & timestamp, LONG64 yearOffset)
 	static_cast<gsoap_eml2_3::_eml23__TimeSeries*>(gsoapProxy2_3)->Time.push_back(ts);
 }
 
-unsigned int TimeSeries::getTimestampIndex(time_t timestamp, LONG64 yearOffset) const
+uint64_t TimeSeries::getTimestampIndex(time_t timestamp, LONG64 yearOffset) const
 {
 	gsoap_eml2_3::_eml23__TimeSeries* timeSeries = static_cast<gsoap_eml2_3::_eml23__TimeSeries*>(gsoapProxy2_3);
 
@@ -65,7 +65,7 @@ unsigned int TimeSeries::getTimestampIndex(time_t timestamp, LONG64 yearOffset) 
 	throw out_of_range("The timestamp has not been found in the allowed range.");
 }
 
-unsigned int TimeSeries::getTimestampIndex(const tm & timestamp, LONG64 yearOffset) const
+uint64_t TimeSeries::getTimestampIndex(const tm & timestamp, LONG64 yearOffset) const
 {
 	gsoap_eml2_3::_eml23__TimeSeries* timeSeries = static_cast<gsoap_eml2_3::_eml23__TimeSeries*>(gsoapProxy2_3);
 
@@ -85,12 +85,12 @@ unsigned int TimeSeries::getTimestampIndex(const tm & timestamp, LONG64 yearOffs
 	throw out_of_range("The timestamp has not been found in the allowed range.");
 }
 
-unsigned int TimeSeries::getTimestampCount() const
+uint64_t TimeSeries::getTimestampCount() const
 {
 	return static_cast<gsoap_eml2_3::_eml23__TimeSeries*>(gsoapProxy2_3)->Time.size();
 }
 
-tm TimeSeries::getTimestampAsTimeStructure(unsigned int index) const
+tm TimeSeries::getTimestampAsTimeStructure(uint64_t index) const
 {
 	gsoap_eml2_3::_eml23__TimeSeries* timeSeries = static_cast<gsoap_eml2_3::_eml23__TimeSeries*>(gsoapProxy2_3);
